flask_app/requirements.txt
Authlib==1.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
